requirements.txt
aiohttp==0.18.4
alembic==1.5.4
attrs==21.2.0
autopep8==1.5.5
cchardet==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
httpie==0.9.2
idna==2.10
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.4
MarkupSafe==2.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.10.0
pycodestyle==2.6.0
Pygments==2.12.0
PyJWT==2.0.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.2.2
datetime==4.5
uuid

